requirements.txt
polars>=0.20.0
dbfread>=2.0.7
pydantic>=2.5.0
loguru>=0.7.2
PySide6>=6.6.0
python-dateutil>=2.8.2
office365-rest-python-client>=2.5.0
pydrive2>=1.17.0
google-api-python-client>=2.110.0
pandas>=2.1.0